Moving leaderboard to scoreboards package
