+++ pkg/scoreboards/leaderboard.go
@@ -1,9 +1,11 @@
-package users
+package scoreboards
 
 import (
 	"fmt"
 	"strconv"
 	"strings"
+
+	"github.com/adanalife/tripbot/pkg/users"
 
 	c "github.com/adanalife/tripbot/pkg/config/tripbot"
 	"github.com/adanalife/tripbot/pkg/database"
@@ -18,7 +20,7 @@ var maxLeaderboardSize = 50
 
 // Leaderboard creates a leaderboard
 func InitLeaderboard() {
-	users := []User{}
+	topUsers := []users.User{}
 
 	ignoredUsers := append(c.IgnoredUsers, strings.ToLower(c.Conf.ChannelName))
 	// we use MySQL-style ? bindvars instead of postgres ones here
@@ -29,12 +31,12 @@ func InitLeaderboard() {
 		terrors.Log(err, "error generating query")
 	}
 	query = database.Connection().Rebind(query)
-	err = database.Connection().Select(&users, query, args...)
+	err = database.Connection().Select(&topUsers, query, args...)
 	if err != nil {
 		terrors.Log(err, "error generating query")
 	}
 
-	for _, user := range users {
+	for _, user := range topUsers {
 		miles := fmt.Sprintf("%.1f", user.Miles)
 		pair := []string{user.Username, miles}
 		Leaderboard = append(Leaderboard, pair)
@@ -42,7 +44,7 @@ func InitLeaderboard() {
 }
 
 func UpdateLeaderboard() {
-	for _, user := range LoggedIn {
+	for _, user := range users.LoggedIn {
 		// skip adding this user if they're a bot or ignored
 		if user.IsBot || c.UserIsIgnored(user.Username) || c.UserIsAdmin(user.Username) {
 			continue
@@ -65,7 +67,7 @@ func strToFloat32(str string) float32 {
 	return float32(value)
 }
 
-func insertIntoLeaderboard(user User) {
+func insertIntoLeaderboard(user users.User) {
 	// first we remove this user from the board
 	removeFromLeaderboard(user.Username)
